third_party/blink/web_tests/http/tests/devtools/components/widget-events-expected.txt

This tests that events are properly propagated through Widget hierarchy.


Running: testShowWidget
Widget()
Widget.show()
  Widget.wasShown()
  Widget.onResize()
Widget.detach()
  Widget.willHide()

Running: testAppendViaDOM
Widget()
Error: Attempt to add widget via regular DOM operation.

Running: testInsertViaDOM
Widget()
Error: Attempt to add widget via regular DOM operation.

Running: testAttachToOrphanNode
Widget()
Widget.show()
Error: Attempt to attach widget to orphan node

Running: testImmediateParent
Parent()
Child()
Child.show()
OK

Running: testDistantParent
Parent()
Child()
Child.show()
OK

Running: testEvents
Parent()
Child()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
Parent.doResize()
Child.show()
  Child.wasShown()
  Child.onResize()
Parent.doResize()
  Child.onResize()
Parent.detach()
  Child.willHide()
  Parent.willHide()
Parent.show()
  Parent.wasShown()
  Child.wasShown()
  Parent.onResize()
  Child.onResize()
Child.detach()
  Child.willHide()
Parent.detach()
  Parent.willHide()

Running: testEventsHideOnDetach
Parent()
Child()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
Parent.doResize()
Child.show()
  Child.wasShown()
  Child.onResize()
Parent.doResize()
  Child.onResize()
Parent.detach()
  Child.willHide()
  Parent.willHide()
Parent.show()
  Parent.wasShown()
  Child.wasShown()
  Parent.onResize()
  Child.onResize()
Child.detach()
  Child.willHide()
Parent.detach()
  Parent.willHide()

Running: testWidgetCounter
Parent()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
Child()
Child.show()
  Child.wasShown()
  Child.onResize()
  widget counter: 1
Child 2()
Child 2.show()
  Child 2.wasShown()
  Child 2.onResize()
  widget counter: 2
Child.detach()
  Child.willHide()
  widget counter: 1
Child 2.detach()
  Child 2.willHide()
  widget counter: 0

Running: testRemoveChild
Parent()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
Child()
Child.show()
  Child.wasShown()
  Child.onResize()
Error: Attempt to remove element containing widget via regular DOM operation

Running: testImplicitRemoveChild
Parent()
Child()
Child.show()
Error: Attempt to remove element containing widget via regular DOM operation

Running: testRemoveChildren
Parent()
Child()
Child.show()
Error: Attempt to remove element containing widget via regular DOM operation

Running: testImplicitRemoveChildren
Parent()
Child()
Child.show()
Error: Attempt to remove element containing widget via regular DOM operation

Running: testShowOnWasShown
Parent()
Child()
Parent.show()
  Parent.wasShown()
Child.show()
  Child.wasShown()
  Parent.onResize()
  Child.onResize()
Parent.detach()
  Child.willHide()
  Parent.willHide()

Running: testShowNestedOnWasShown
Top()
Middle()
Bottom()
Middle.show()
Top.show()
  Top.wasShown()
Bottom.show()
  Middle.wasShown()
  Bottom.wasShown()
  Top.onResize()
  Middle.onResize()
  Bottom.onResize()
Top.detach()
  Bottom.willHide()
  Middle.willHide()
  Top.willHide()

Running: testDetachOnWasShown
Parent()
Child()
Child.show()
Parent.show()
  Parent.wasShown()
Child.detach()
  Parent.onResize()
Parent.detach()
  Parent.willHide()

Running: testShowOnWillHide
Parent()
Child()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
Child.show()
  Child.wasShown()
  Child.onResize()
Parent.detach()
  Child.willHide()
  Parent.willHide()
Child.show()

Running: testDetachOnWillHide
Parent()
Child()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
Child.show()
  Child.wasShown()
  Child.onResize()
Parent.detach()
  Child.willHide()
  Parent.willHide()
Child.detach()

Running: testShowDetachesFromPrevious
Parent1()
Parent2()
Child()
Parent1.show()
  Parent1.wasShown()
  Parent1.onResize()
Parent2.show()
  Parent2.wasShown()
  Parent2.onResize()
Child.show()
  Child.wasShown()
  Child.onResize()
Child.show()
Child.detach()
  Child.willHide()
  Child.wasShown()
  Child.onResize()

Running: testResizeOnWasShown
Parent()
Child()
Child.show()
Parent.show()
  Parent.wasShown()
Child.doResize()
  Child.wasShown()
  Parent.onResize()
  Child.onResize()
Parent.detach()
  Child.willHide()
  Parent.willHide()

Running: testReparentWithinWidget
Parent()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
Child()
Child.show()
  Child.wasShown()
  Child.onResize()
Child.show()
  Child.onResize()

Running: testDetachChildWidgetsRemovesHiddenChildren
Parent()
visibleChild()
hiddenChild()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
visibleChild.show()
  visibleChild.wasShown()
  visibleChild.onResize()
hiddenChild.show()
  hiddenChild.wasShown()
  hiddenChild.onResize()
  hiddenChild.willHide()
visibleChild.detach()
  visibleChild.willHide()
hiddenChild.detach()
Parent element has 0 child elements
